+++ @types/@material-ui/index.d.ts
@@ -1,0 +1,3 @@
+declare module '@material-ui/utils' {
+  export function deepmerge(target: unknown, source: unknown): object
+}