+++ packages/picasso/src/Icon/Drag24.tsx
@@ -1,0 +1,52 @@
+import React, { forwardRef, Ref } from 'react'
+import cx from 'classnames'
+import { withStyles } from '@material-ui/core/styles'
+import { StandardProps } from '@toptal/picasso-shared'
+
+import kebabToCamelCase from '../utils/kebab-to-camel-case'
+import styles from './styles'
+const BASE_SIZE = 24
+type ScaleType = 1 | 2 | 3 | 4
+export interface Props extends StandardProps {
+  scale?: ScaleType
+  color?: string
+  base?: number
+}
+const SvgDrag24 = forwardRef(function SvgDrag24(
+  props: Props,
+  ref: Ref<SVGSVGElement>
+) {
+  const {
+    classes: availableClasses,
+    className,
+    style = {},
+    color,
+    scale,
+    base
+  } = props
+  const classes = [availableClasses.root, className]
+  const scaledSize = base || BASE_SIZE * Math.ceil(scale || 1)
+  const colorClassName = kebabToCamelCase(`${color}`)
+
+  if (availableClasses[colorClassName]) {
+    classes.push(availableClasses[colorClassName])
+  }
+
+  const svgStyle = {
+    minWidth: `${scaledSize}px`,
+    minHeight: `${scaledSize}px`,
+    ...style
+  }
+  return (
+    <svg
+      viewBox='0 0 24 24'
+      className={cx(...classes)}
+      style={svgStyle}
+      ref={ref}
+    >
+      <path d='M6 20a1 1 0 110 2 1 1 0 010-2zm6 0a1 1 0 110 2 1 1 0 010-2zm6 0a1 1 0 110 2 1 1 0 010-2zM6 14a1 1 0 110 2 1 1 0 010-2zm6 0a1 1 0 110 2 1 1 0 010-2zm6 0a1 1 0 110 2 1 1 0 010-2zM6 8a1 1 0 110 2 1 1 0 010-2zm6 0a1 1 0 110 2 1 1 0 010-2zm6 0a1 1 0 110 2 1 1 0 010-2zM6 2a1 1 0 110 2 1 1 0 010-2zm6 0a1 1 0 110 2 1 1 0 010-2zm6 0a1 1 0 110 2 1 1 0 010-2z' />
+    </svg>
+  )
+})
+SvgDrag24.displayName = 'SvgDrag24'
+export default withStyles(styles)(SvgDrag24)